refactor: rename and move functions to new files
