+++ src/types/typeGuards/index.ts
@@ -1,3 +1,5 @@
 export * from "./compound";
+export * from "./intrinsic";
+export * from "./literal";
 export * from "./objects";
 export * from "./single";
